feat: createLecture과 createCurriculum에서 쓰는 함수들 훅으로 분리 및 난이도 값 수정
